pyproject.toml
[tool.poetry]
name = "drf-coursework"
version = "0.1.0"
description = ""
authors = ["RocketMenace <grizzlygit777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
celery = "^5.4.0"
redis = "^5.2.0"
django-celery-beat = "^2.7.0"
django-environ = "^0.11.2"
black = "^24.10.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.28.0"
coverage = "^7.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
